+++ src/constants.js
@@ -5,7 +5,7 @@
  */
 
 /** Pulse Card version string. */
-export const VERSION = '0.3.2';
+export const VERSION = '0.4.0';
 
 /** Console log prefix. */
 export const LOG_PREFIX = 'Pulse Card:';